Added models and connected services and repositories for post, comment, file and role

+++ WorkAndTravel/src/main/java/com/diplomska/backend/service/implementation/UserServiceImpl.java
@@ -28,7 +28,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public User findById(Integer id) {
+    public User findById(Long id) {
         return this.userRepository.findById(id).orElseThrow(UserNotFoundException::new);
     }
