+++ WorkAndTravel/src/main/java/com/diplomska/backend/service/interfaces/UserService.java
@@ -7,6 +7,6 @@ import java.util.List;
 public interface UserService {
     User create (User user);
     User update (User user);
-    User findById (Integer id);
+    User findById (Long id);
     List<User> findAll();
 }
